Proxy: forward http headers from injector to client
